refactor: update create pwd
