+++ app/src/main/java/com/cj/bunnywallet/Constant.kt
@@ -2,6 +2,7 @@ package com.cj.bunnywallet
 
 // datastore key
 const val KEY_PWD = "pwd"
+const val KEY_ENABLE_BIOMETRICS = "enable_biometrics"
 const val KEY_MNEMONIC = "mnemonic"
 const val KEY_HINT = "hint"
